@@ -100,10 +100,9 @@
     "@popperjs/core": "2.11.2",
     "@types/bootstrap": "^5.1.9",
     "@types/jquery": "^3.5.14",
-    "@wisemapping/editor": "^1.4.1",
     "ajv": "^8.11.0",
     "bootstrap": "5.1.3",
-    "browser-sync-client": "^3.0.2",
+    "browser-sync-client": "^2.27.10",
     "dayjs": "1.10.7",
     "material-design-icons": "^3.0.1",
     "ngx-infinite-scroll": "10.0.1",
@@ -111,14 +110,14 @@
     "ngx-xml2json": "^1.0.2",
     "primeicons": "^5.0.0",
     "primeng": "file:../primeng/dist/primeng-28.7.0.tgz",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-intl": "5.24.8",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-intl": "6.6.5",
     "registry-url": "^6.0.1",
     "rxjs": "7.5.4",
     "swagger-ui-dist": "4.5.2",
     "tslib": "2.3.1",
-    "typescript": "^4.7.4",
+    "typescript": "^4.7.3",
     "vkbeautify": "^0.99.3",
     "xml-formatter": "^2.6.1",
     "xml2js": "^0.4.23",
@@ -141,12 +140,16 @@
     "@babel/preset-typescript": "^7.16.7",
     "@types/jest": "27.4.1",
     "@types/node": "16.11.26",
-    "@types/react-dom": "^18.0.0",
     "@types/react-intl": "^3.0.0",
     "@typescript-eslint/eslint-plugin": "6.21.0",
     "@typescript-eslint/parser": "6.21.0",
+    "@wisemapping/core-js": "file:../createyourhumanity-mindmap/packages/core-js",
+    "@wisemapping/editor": "file:../createyourhumanity-mindmap/packages/editor",
+    "@wisemapping/mindplot": "file:../createyourhumanity-mindmap/packages/mindplot",
+    "@wisemapping/web2d": "file:../createyourhumanity-mindmap/packages/web2d",
     "babel-loader": "^8.0.6",
-    "browser-sync": "2.27.7",
+    "bootstrap-loader": "^4.0.4",
+    "browser-sync": "^2.29.3",
     "browser-sync-webpack-plugin": "2.3.0",
     "concurrently": "7.0.0",
     "copy-webpack-plugin": "10.2.4",
